Implement text search container for displaying pure text search result list view.
